Move profile components to components folder
