+++ frontend/src/components/profile/ActivityMap.scss
