+++ frontend/src/components/profile/ActivityMap.tsx
