+++ frontend/src/components/profile/ProfileHeader.scss
