+++ frontend/src/components/profile/UserStats.tsx
